completed(...) now deallocates a callback immediately after invocation.
This is necessary to avoid a tricky deadlock that can occur when, just
after invocation, we acquire the final reference to a file.

If the callback is not deallocated before we drop this reference, we'd
wind up with a deadlock as the callback could reference an activity on
the file that is about to be destroyed.

Of course, the file would never be destroyed because it would wait for
any pending activities to complete first.